Revert "Added GlobalVelocity and HeadingVelocity model"
