+++ test.py
@@ -4,30 +4,24 @@ import unittest
 class TestExceptions(unittest.TestCase):
     def test_invalidPNGainInput(self):
         with self.assertRaises(PN.InvalidProportionalGainError):
-            PN.PN(PN.HeadingVelocity(0,0,0,0),PN.GlobalVelocity(0,0,0,0),0)
+            PN.PN(PN.Body(0,0,0,0),PN.Body(0,0,0,0),0)
 
         with self.assertRaises(TypeError):
-            PN.PN(PN.HeadingVelocity(0,0,0,0),PN.GlobalVelocity(0,0,0,0),"str")
+            PN.PN(PN.Body(0,0,0,0),PN.Body(0,0,0,0),"str")
 
-    def test_invalidHeadingVelocityBody(self):
+    def test_invalidBodies(self):
         with self.assertRaises(TypeError):
-            PN.PN("str",PN.HeadingVelocity(0,0,0,0),1)
+            PN.PN("str",PN.Body(0,0,0,0),1)
         with self.assertRaises(TypeError):
-            PN.PN(PN.HeadingVelocity(0,0,0,0),"str",1)
-
-    def test_invalidGlobalVelocityBody(self):
-        with self.assertRaises(TypeError):
-            PN.PN("str",PN.GlobalVelocity(0,0,0,0),1)
-        with self.assertRaises(TypeError):
-            PN.PN(PN.GlobalVelocity(0,0,0,0),"str",1)
+            PN.PN(PN.Body(0,0,0,0),"str",1)
     
     def test_invalidPNOptions(self):
         with self.assertRaises(TypeError):
-            PN.PN(PN.HeadingVelocity(0,0,0,0),PN.GlobalVelocity(0,0,0,0),1,options="str")
+            PN.PN(PN.Body(0,0,0,0),PN.Body(0,0,0,0),1,options="str")
 
     def test_sameCoords(self):
-        with self.assertRaises(PN.exceptions.OutOfBoundsRangeError):
-            PN.PN(PN.HeadingVelocity(0,0,0,0),PN.GlobalVelocity(0,0,0,0)).calculate()
+        with self.assertRaises(PN.OutOfBoundsRangeError):
+            PN.PN(PN.Body(0,0,0,0),PN.Body(0,0,0,0)).calculate()
         
 
 class TestPNOptions(unittest.TestCase):
@@ -44,7 +38,7 @@ class TestPNOptions(unittest.TestCase):
     
     def test_optionsWork(self):
         opt1 = PN.PNOptions(return_R=True,return_Rdot=True,return_Vc=True,return_lambda=True,return_lambdad=True)
-        ret1 = PN.PN(PN.HeadingVelocity(10,1,0,0),PN.GlobalVelocity(0,10,10,0),options=opt1).calculate()
+        ret1 = PN.PN(PN.Body(10,1,0,0),PN.Body(0,10,10,0),options=opt1).calculate()
         self.assertTrue("R" in ret1.keys())
         self.assertTrue("Rdot" in ret1.keys())
         self.assertTrue("Vc" in ret1.keys())
@@ -53,7 +47,7 @@ class TestPNOptions(unittest.TestCase):
         self.assertTrue("nL" in ret1.keys())    
 
         opt2 = PN.PNOptions(return_R=True,return_Rdot=True,return_Vc=False,return_lambda=True,return_lambdad=False)
-        ret2 = PN.PN(PN.HeadingVelocity(10,1,0,0),PN.GlobalVelocity(0,10,10,0),options=opt2).calculate()
+        ret2 = PN.PN(PN.Body(10,1,0,0),PN.Body(0,10,10,0),options=opt2).calculate()
         self.assertTrue("R" in ret2.keys())
         self.assertTrue("Rdot" in ret2.keys())
         self.assertFalse("Vc" in ret2.keys())
@@ -61,7 +55,7 @@ class TestPNOptions(unittest.TestCase):
         self.assertFalse("lambdad" in ret2.keys())
         self.assertTrue("nL" in ret2.keys())
 
-        ret3 = PN.PN(PN.HeadingVelocity(10,1,0,0),PN.GlobalVelocity(0,10,10,0)).calculate()
+        ret3 = PN.PN(PN.Body(10,1,0,0),PN.Body(0,10,10,0)).calculate()
         self.assertIsInstance(ret3,float)
 
 if __name__ == '__main__':
